pyproject.toml
[tool.poetry]
name = "akuire"
version = "0.1.1"
description = ""
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]
readme = "README.md"
packages = [{ include = "akuire" }]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.5"
koil = "^0.3.6"


[tool.poetry.group.dev.dependencies]
pytest-aiohttp = "^1.0.5"
numpy = "^2.0.0"
numba = "^0.60.0"
scipy = "^1.13.1"
opencv-python-headless = "^4.10.0.84"
nanoimagingpack = "^2.1.3.dev1"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
akuire = "akuire.__main__:main"

[tool.pytest.ini_options]
markers = [
    "composition: test composition",
    "external: test external",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
